Changed view on Frontend

+++ database/migrations/2022_03_25_194511_add_user_id_to_applyleaves_table.php
@@ -14,7 +14,7 @@ class AddUserIdToApplyleavesTable extends Migration
     public function up()
     {
         Schema::table('applyleaves', function (Blueprint $table) {
-            $table->unsignedInteger('user_id')->references('id')->on('users');
+            $table->unsignedBigInteger('user_id')->references('id')->on('users');
         });
     }
